Add static content to Personal Code view
- add static page content using localisation
- use static placeholder for dynamic text '[Selected PSC]'
- refactor Skeleton Four as Personal Code
- rename files accordingly
- alter previous view's Continue button target
