+++ locales/cy/global.json
@@ -3,5 +3,6 @@
     "global_banner_feedback_link": "adborth",
     "global_banner_help_improve": " yn ein helpu i'w wella.",
     "global_confirm_continue_button": "Cadarnhau a parhau",
+    "global_save_continue_button": "Cadw a parhau",
     "global_continue_button": "Parhau"
 }
