+++ locales/cy/personal_code.json
@@ -1,0 +1,5 @@
+{
+  "personal_code_title": "to be translated",
+  "personal_code_details_summaryText": "to be translated",
+  "personal_code_details_text": "to be translated"
+}
