+++ locales/en/global.json
@@ -3,5 +3,6 @@
     "global_banner_feedback_link": "feedback",
     "global_banner_help_improve": " will help us to improve it.",
     "global_confirm_continue_button": "Confirm and continue",
+    "global_save_continue_button": "Save and continue",
     "global_continue_button": "Continue"
 }
