+++ locales/en/personal_code.json
@@ -1,0 +1,5 @@
+{
+  "personal_code_title": "What is their Companies House personal code?",
+  "personal_code_details_summaryText": "Where to find the Companies House personal code",
+  "personal_code_details_text": "There is an 11 character code that is given to a person after they have verified their identity with Companies House."
+}
