+++ src/routers/__utils/index.ts
@@ -3,7 +3,7 @@ import CompanyNumberRouter from "./../companyNumberRouter";
 import ConfirmCompanyRouter from "./../confirmCompanyRouter";
 import PscTypeRouter from "./../psc.type.router";
 import SkeletonThreeRouter from "./../skeleton.three.router";
-import SkeletonFourRouter from "./../skeleton.four.router";
+import PersonalCodeRouter from "../personalCode.router";
 import SkeletonFiveRouter from "./../skeleton.five.router";
 import PscVerifiedRouter from "./../pscVerifiedRouter";
-export { StartRouter, CompanyNumberRouter, ConfirmCompanyRouter, PscTypeRouter, SkeletonThreeRouter, SkeletonFourRouter, SkeletonFiveRouter, PscVerifiedRouter };
+export { StartRouter, CompanyNumberRouter, ConfirmCompanyRouter, PscTypeRouter, SkeletonThreeRouter, PersonalCodeRouter, SkeletonFiveRouter, PscVerifiedRouter };
