+++ src/routers/personalCode.router.ts
@@ -1,11 +1,12 @@
 import { NextFunction, Request, Response, Router } from "express";
-import { SkeletonFourHandler } from "./handlers/skeleton_four";
+import { PersonalCodeHandler } from "./handlers/personal_code/personalCode";
 import { handleExceptions } from "../utils/async.handler";
 import logger from "../lib/Logger";
+
 const router: Router = Router();
 
 router.get("/", handleExceptions(async (req: Request, res: Response) => {
-    const handler = new SkeletonFourHandler();
+    const handler = new PersonalCodeHandler();
     const { templatePath, viewData } = handler.executeGet(req, res);
     res.render(templatePath, viewData);
 }));
