+++ src/views/router_views/personal_code/personal_code.njk
@@ -1,0 +1,38 @@
+{% extends "layouts/default.njk" %}
+
+{% from "govuk/components/input/macro.njk" import govukInput %}
+{% from "govuk/components/details/macro.njk" import govukDetails %}
+{% from "govuk/components/button/macro.njk" import govukButton %}
+
+{% set title = i18n.personal_code_title %}
+
+{% block main_content %}
+
+  <span class="govuk-caption-xl">
+    [Selected PSC]
+  </span>
+  <h1 class="govuk-heading-l" id="personal-code--title">
+      {{title}}
+  </h1>
+    <div class="govuk-form-group">
+      {{ govukInput({
+        id: "personalCode",
+        classes: "govuk-input--width-20",
+        name: "personalCode",
+        value: ""
+      }) }}
+      {{ govukDetails({
+        summaryText: i18n.personal_code_details_summaryText,
+        text: i18n.personal_code_details_text,
+        classes: "govuk-!-padding-top-6"
+      }) }}
+      {{ govukButton({
+          text: i18n.global_save_continue_button,
+          href: Urls.SKELETON_FIVE,
+          classes: "govuk-button--start govuk-!-margin-top-2 govuk-!-margin-bottom-8",
+          isStartButton: true
+      }) }}
+    </div>
+
+
+{% endblock %}
