+++ src/views/router_views/skeleton_three/skeleton_three.njk
@@ -9,8 +9,8 @@
   </h1>
 
       {{ govukButton({
-          text: "Skeleton Four",
-          href: Urls.SKELETON_FOUR,
+          text: i18n.global_continue_button,
+          href: Urls.PERSONAL_CODE,
           classes: "govuk-button--start govuk-!-margin-top-2 govuk-!-margin-bottom-8",
           isStartButton: true
       }) }}
